[query] Move typeToTypeInfo and defaultValue to ptypes package
