+++ hail/src/main/scala/is/hail/types/physical/PPrimitive.scala
@@ -9,6 +9,8 @@ import is.hail.utils._
 trait PPrimitive extends PType {
   def byteSize: Long
 
+  def _construct(mb: EmitMethodBuilder[_], region: Value[Region], pc: PCode): PCode = pc
+
   def _copyFromAddress(region: Region, srcPType: PType, srcAddress: Long, deepCopy: Boolean): Long = {
     if (!deepCopy)
       return srcAddress
