+++ hail/src/main/scala/is/hail/types/physical/PType.scala
@@ -459,7 +459,7 @@ abstract class PType extends Serializable with Requiredness {
 
   def deepRename(t: Type): PType = this
 
-  def defaultValue: PCode = PCode(this, ir.defaultValue(this))
+  def defaultValue: PCode = PCode(this, is.hail.types.physical.defaultValue(this))
 
 
   final def typeCheck(a: Any): Boolean = a == null || _typeCheck(a)
